rearranged files and folders. Changed class xxxx extends to functions. Created a loop for items, will need one for departments too.
